removed hook, added component for xss protection
