--- frontend/src/features/map_planning/layers/plant/hooks/useExtendedPlantsSummary.ts
@@ -1,10 +1,0 @@
-import { PlantsSummaryDto } from '@/bindings/definitions';
-import { ExtendedPlantsSummary } from '@/utils/extendedPlantsSummary';
-import { useMemo } from 'react';
-
-export function useExtendedPlantSummaryTransformed(plant: PlantsSummaryDto) {
-  const extendedPlantSummaryTransformed = useMemo(() => {
-    return new ExtendedPlantsSummary(plant);
-  }, [plant]);
-  return extendedPlantSummaryTransformed;
-}
